refactor(watt): inline template inside component
